Initial working lambda support
