+++ go.mod
@@ -3,17 +3,13 @@ module github.com/wspowell/spiderweb
 go 1.16
 
 require (
-	github.com/andybalholm/brotli v1.0.1 // indirect
-	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/aws/aws-lambda-go v1.13.3
 	github.com/fasthttp/router v1.3.10
 	github.com/google/gofuzz v1.2.0
 	github.com/klauspost/compress v1.11.13 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/savsgio/gotils v0.0.0-20210316171653-c54912823645 // indirect
-	github.com/sirupsen/logrus v1.8.1 // indirect
 	github.com/stretchr/objx v0.3.0 // indirect
 	github.com/stretchr/testify v1.7.0
-	github.com/valyala/bytebufferpool v1.0.0 // indirect
 	github.com/valyala/fasthttp v1.23.0
 	github.com/wspowell/errors v0.1.0
 	github.com/wspowell/local v0.0.3
